backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
celery
redis
python-dotenv
openai
pinecone
Pillow